1. add eventbus library in build.gradle; 2. add internet permission , access network state permission and provider in AndroidManifest.xml; 3. add RetrofitDataAgent at 4. change code in DestinationDetailActivity.java 5. change code in HomeActivity.java 6. change DestinationVo list name , return mDestinationList.size () in getCount( ) , add setNewdata( ) at DestinationAdapter.java; 7. add onEventMainThread(Object obj) at BaseModel.java; 8. add FestivalModel.java; 9. add HotelModel.java; 10.add PackageModel.java; 11. add api for hotels,festivals,pacakges at DestinationApi.java; 12. add loadFestivals() , loadHotels(),loadPackages() at DestinationDataAgent.java; 13. add RetrofitDataAgent.java; 14. add DestinationContract.java; 15. add DestinationDBHelper.java; 16. add DestinationProvider.java; 17. add FestivalListResponse.java; 18. add HotelListResponse.java; 19. add PackageListResponse.java; 20. change placeId from int to long at AttractionPlacesVo.java; 21. change cityId from int to long at CityVo.java; 22. change id from int to long at DestinationVO.java; 23. change totalDays from int to String 24. change stateId from int to long; 25. add DataEvent.java;

+++ PADC-FP-GoldenMyanmarTour/app/src/main/java/com/padc/goldenmyanmartour/data/vo/AttractionPlacesVO.java
@@ -7,7 +7,7 @@ import com.google.gson.annotations.SerializedName;
  */
 public class AttractionPlacesVO {
     @SerializedName("place-id")
-    public int placeId;
+    public long placeId;
     @SerializedName("title")
     public String title;
     @SerializedName("image")
@@ -17,11 +17,11 @@ public class AttractionPlacesVO {
     @SerializedName("note-to-visitor")
     public String noteToVisitor;
 
-    public int getPlaceId() {
+    public long getPlaceId() {
         return placeId;
     }
 
-    public void setPlaceId(int placeId) {
+    public void setPlaceId(long placeId) {
         this.placeId = placeId;
     }
